reorganized components into folders. passed calculateMacro function to data component
