--- src/components/MacroImportant.js
@@ -1,18 +1,0 @@
-import React from 'react';
-
-function MacroImportant(){
-       return (
-        <div className=" container row my-5 border-bottom border-warning">
-            <h1 className="display-4 mb-3">Why are Counting Macros so Important?</h1>
-            <p className="lead mx-auto">
-                Getting the correct amount of protein, carbs and fat into your macro diet will help you reach your fitness goals, whether you want to lose weight or put on some muscle. Meal plans should have a balance of all the three key macronutrients: protein, carbs, and fats.
-            </p>
-            <p className="lead mx-auto">    
-                When your diet is too low in protein, you could lose muscle instead of fat, which can lead to a slower metabolism.
-                If you aren't getting enough carbs, you may be very low in energy. If you are low in fats, you might suffer from hormonal imbalances. If carbs or fats are too high, you may not see fat loss results.
-            </p>
-        </div>
-       );
-}
-
-export default MacroImportant;
