+++ src/components/pages/Data.js
@@ -1,13 +1,13 @@
 import React from 'react';
-import MacroForm from './MacroForm';
+import MacroForm from '../layout/MacroForm';
 
 
 function Data(props) {
         return (
            <div>
-                <form onSubmit={props.calculateMacros} className="row col-xl-8 col-lg-6 col-md-8 col-sm-10 mx-auto border border-dark border-bottom-0 border-right-0 border-left-0 pt-5">
+                <form onSubmit={props.calculateMacros} className="row col-xl-12 col-lg-12 col-md-8 col-sm-10 mx-auto border border-dark border-bottom-0 border-right-0 border-left-0 pt-5">
                     <h1 className="display-4 text-center col-xl-12">What Did You Eat Today?</h1>
-                    <div className="form-group col-xl-4">
+                    <div className="form-group col-xl-4 col-lg-4 col-md-6">
                         <label>Day</label>
                         <select className="custom-select" name="day" id="daySelect">
                             <option defaultValue>Choose...</option>
@@ -20,7 +20,7 @@ function Data(props) {
                             <option value="Saturday">Saturday</option>
                         </select>
                     </div>
-                    <div className="form-group col-xl-4">
+                    <div className="form-group col-xl-4 col-lg-4 col-md-6">
                         <label>Type</label>
                         <select className="custom-select" name="type" id="foodType">
                             <option defaultValue>Choose...</option>
@@ -31,19 +31,19 @@ function Data(props) {
                             <option value="Snack">Snack</option>
                         </select>
                     </div>
-                    <div className="form-group col-xl-4" name="foodName">
+                    <div className="form-group col-xl-4 col-lg-4 col-md-6" name="foodName">
                         <label>Food Name</label>
                         <input type="text" className="form-control form1" id="foodName" aria-describedby="emailHelp" placeholder="Enter Food Name Here" />
                     </div>
-                    <div className="form-group col-xl-4" name="protein">
+                    <div className="form-group col-xl-4 col-lg-4 col-md-6" name="protein">
                         <label>Protein</label>
                         <input type="text" className="form-control form2" id="proteinCount" aria-describedby="emailHelp" placeholder="Enter Protein Intake" />                              
                     </div>
-                    <div className="form-group col-xl-4" name="carbs">
+                    <div className="form-group col-xl-4 col-lg-4 col-md-6" name="carbs">
                         <label>Carbs</label>
                         <input type="text" className="form-control form3" id="carbCount" placeholder="Enter Carb Intake"/>
                     </div>
-                    <div className="form-group col-xl-4" name="fat">
+                    <div className="form-group col-xl-4 col-lg-4 col-md-6" name="fat">
                         <label>Fat</label>
                         <input type="text" className="form-control form4" id="fatCount" placeholder="Enter Fat Intake"/>
                     </div>
