+++ src/components/pages/Landing.js
@@ -1,9 +1,6 @@
 import React from 'react';
-import '../App.css';
+import '../../App.css';
 import { Link } from 'react-router-dom';
-
-
-
 
 function Landing(){
         return (
